+++ Modules/GuildQuotes.lua
@@ -12,10 +12,6 @@ Dadabase.DatabaseManager:RegisterModule("guildquotes", {
     dbVersion = 1,
     defaultSettings = {
         enabled = false,
-        triggers = {
-            wipe = false,
-            death = false
-        },
         groups = {
             raid = false,
             party = false